Proposing MIT License
Reason for proposing license: To make it more clear legal wise for anybody wanting to use this in a project in the future. The readme is pretty clear but having a proper license may be better in some cases.

It also doesn't hurt to add one, even if it isn't the MIT license.